Short timezones and link to month list

+++ assets/js/timezone-converter.js
@@ -5,7 +5,7 @@
         // Convert to local browser time for actual display.
         const date = new Date(Date.parse(timeEl.dateTime));
         const options = {
-            timeZoneName: 'long',
+            timeZoneName: 'short',
             year: 'numeric',
             month: 'long',
             day: 'numeric',
